+++ vmdb/app/helpers/ems_infra_helper/textual_summary.rb
@@ -40,7 +40,7 @@ module EmsInfraHelper::TextualSummary
   end
 
   def textual_port
-    ["openstack", "rhevm"].include?(@ems.emstype) ? {:label => "API Port", :value => @ems.port} : nil
+    ["openstack", "openstack_infra", "rhevm"].include?(@ems.emstype) ? {:label => "API Port", :value => @ems.port} : nil
   end
 
   def textual_cpu_resources
