+++ vmdb/app/views/layouts/_multi_auth_credentials.html.erb
@@ -13,7 +13,7 @@
       <% if ["ems_cloud","ems_infra"].include?(params[:controller]) %><%# Special tabs for EMS edit %>
         <% display = @edit[:new][:emstype] == "rhevm" ? "" : "none" %>
         <li id="metrics_li" style="display:<%= display %>;"><a href="#metrics">C & U Database</a></li>
-        <% display = @edit[:new][:emstype] == "openstack" ? "" : "none" %>
+        <% display = ["openstack", "openstack_infra"].include?(@edit[:new][:emstype]) ? "" : "none" %>
         <li id="amqp_li" style="display:<%= display %>;"><a href="#amqp">AMQP</a></li>
       <% else %>
         <li><a href="#remote">Remote Login</a></li>
